requirements.txt
# Automação e controle do Moodle
pyautogui
mss
selenium
webdriver-manager
python-dotenv

# Dashboard e visualização
streamlit
plotly
